requirements.txt
dependency-injector==4.46.0
fastapi==0.115.8
pydantic==2.10.6
pydantic-settings==2.8.1
pytest==8.3.4
python-dotenv==1.0.1
SQLAlchemy==2.0.38
sqlmodel==0.0.23
starlette==0.44.0
ulid-py==1.1.0
